preparing to test what will be spack updater first release (#34)
* preparing to test what will be spack updater first release

I refactored to simplify workflows, so the composite actions
handle more of the logic. The check releases workflow also now does
not require a VERSION file or repo to be defined, as this can
be derived from the package file. This update will use main branch
for a night or two (to ensure everything is still working) and then I'll
make a proper release of the action and use that tag here!

Signed-off-by: vsoch <vsoch@users.noreply.github.com>